LayoutTests/http/tests/plugins/plugin-document-has-focus-expected.txt

CONSOLE MESSAGE: PLUGIN: keyDown 'A'
CONSOLE MESSAGE: PLUGIN: keyUp 'A'
This tests that a full page plugin gets keyboard focus without the need to first click it.
